refact: move Toot to core.toot
