+++ app/src/main/java/com/jeanbarrossilva/mastodonte/app/MastodonteModule.kt
@@ -1,9 +1,9 @@
 package com.jeanbarrossilva.mastodonte.app
 
 import com.jeanbarrossilva.mastodonte.core.profile.ProfileProvider
-import com.jeanbarrossilva.mastodonte.core.profile.toot.TootProvider
 import com.jeanbarrossilva.mastodonte.core.sample.profile.SampleProfileProvider
-import com.jeanbarrossilva.mastodonte.core.sample.profile.toot.SampleTootProvider
+import com.jeanbarrossilva.mastodonte.core.sample.toot.SampleTootProvider
+import com.jeanbarrossilva.mastodonte.core.toot.TootProvider
 import com.jeanbarrossilva.mastodonte.platform.theme.reactivity.OnBottomAreaAvailabilityChangeListener
 import org.koin.core.module.Module
 import org.koin.dsl.module
