+++ core/sample/src/main/java/com/jeanbarrossilva/mastodonte/core/sample/toot/SampleTootWriter.kt
@@ -1,7 +1,7 @@
-package com.jeanbarrossilva.mastodonte.core.sample.profile.toot
+package com.jeanbarrossilva.mastodonte.core.sample.toot
 
-import com.jeanbarrossilva.mastodonte.core.profile.toot.Toot
 import com.jeanbarrossilva.mastodonte.core.sample.profile.edit.replacingOnceBy
+import com.jeanbarrossilva.mastodonte.core.toot.Toot
 import kotlinx.coroutines.flow.update
 
 /** Performs [Toot]-related writing operations. **/
