+++ core/src/main/java/com/jeanbarrossilva/mastodonte/core/profile/Profile.kt
@@ -1,7 +1,7 @@
 package com.jeanbarrossilva.mastodonte.core.profile
 
-import com.jeanbarrossilva.mastodonte.core.profile.toot.Account
-import com.jeanbarrossilva.mastodonte.core.profile.toot.Toot
+import com.jeanbarrossilva.mastodonte.core.toot.Account
+import com.jeanbarrossilva.mastodonte.core.toot.Toot
 import java.io.Serializable
 import java.net.URL
 import kotlinx.coroutines.flow.Flow
