+++ core/src/main/java/com/jeanbarrossilva/mastodonte/core/toot/Author.kt
@@ -1,4 +1,4 @@
-package com.jeanbarrossilva.mastodonte.core.profile.toot
+package com.jeanbarrossilva.mastodonte.core.toot
 
 import java.io.Serializable
 import java.net.URL
