+++ core/src/test/java/com/jeanbarrossilva/mastodonte/core/toot/TootTests.kt
@@ -1,6 +1,6 @@
-package com.jeanbarrossilva.mastodonte.core.profile.toot
+package com.jeanbarrossilva.mastodonte.core.toot
 
-import com.jeanbarrossilva.mastodonte.core.profile.toot.test.TestToot
+import com.jeanbarrossilva.mastodonte.core.toot.test.TestToot
 import kotlin.test.Test
 import kotlin.test.assertFalse
 import kotlin.test.assertTrue
