+++ platform/ui/src/main/java/com/jeanbarrossilva/mastodonte/platform/ui/AccountFormatter.kt
@@ -1,6 +1,6 @@
 package com.jeanbarrossilva.mastodonte.platform.ui
 
-import com.jeanbarrossilva.mastodonte.core.profile.toot.Account
+import com.jeanbarrossilva.mastodonte.core.toot.Account
 
 /** Deals with [Account]-related formatting. **/
 object AccountFormatter {
